:art: refactor : 한 클래스 내에서만 사용되는 로직 합치기
